Create GeneratePassword.py
Modules Used:
Random Module:
Random module is used to perform the random generations. We are making use of random.sample module here. If you will observe in the output all characters will be unique. random.sample() never repeats characters. If you don’t want to repeat characters or digits in the random string, then use random.sample() but it is less secure because it will reduce the probability of combinations because we are not allowing repetitive letters and digits.

String Module:
The string module contains a number of useful constants, classes and a number of functions to process the standard python string.