Create Phone_converter.py
This program converts user phone number to from digits to words.

This program uses the logic of data structure called dictionary {}. 

The user input is set as the key while each key is assigned their values in words. Therefore when ever a user input any valid key, the assigned value will be given as an output.